Add changes in relation to Anmol's comments (change to genericclioptions location)
